Cleaned up some internationalization strings
