+++ src/main/resources/i18n/messages.properties
@@ -994,7 +994,7 @@ project.settings.coverage.maximum_coverage=Maximum Coverage:
 ProjectMetadataTemplates.title=Templates
 ProjectMetadataTemplates.fields={0} fields
 
-MetadataFields.title=Metadata Fields
+MetadataFields.title=Fields
 
 MetadataFieldsList.create=Create Template
 MetadataFieldsList.empty=No metadata has been added to this project.
@@ -1002,17 +1002,17 @@ MetadataFieldsList.empty=No metadata has been added to this project.
 MetadataField.label=Label
 MetadataField.type=Type
 
-MetadataTemplatesList.delete-confirm=Delete this template?
-MetadataTemplatesList.empty=There are no metadata templates associated with this project.
+MetadataTemplates.delete-confirm=Delete this template?
+MetadataTemplates.empty=There are no metadata templates associated with this project.
 
-MetadataTemplatesList.default=Default
-MetadataTemplatesList.set-as-default=Set as Default
-MetadataTemplatesList.cannot-remove-default=Cannot remove a default template. Please select another template as a default first
-MetadataTemplatesList.remove=Remove
-MetadataTemplatesList.download=Download
+MetadataTemplates.default=Default
+MetadataTemplates.set-as-default=Set as Default
+MetadataTemplates.cannot-remove-default=Cannot remove a default template. Please select another template as a default first
+MetadataTemplates.remove=Remove
+MetadataTemplates.download=Download
 
-MetadataTemplatesList.allFields=All Fields
-MetadataTemplatesList.allFields-description=This template displays all the metadata fields on a project. Note that setting this as the default could have an adverse effect on the linelist page performance.
+MetadataTemplates.allFields=All Fields
+MetadataTemplates.allFields-description=This template displays all the metadata fields on a project. Note that setting this as the default could have an adverse effect on the linelist page performance.
 
 
 CreateMetadataTemplate.title=Create Metadata Template
